fix danh sach san pham
