+++ AppKhachHang/app/src/main/java/com/example/appkhachhang/Fragment/HomeFragment.java
@@ -1,7 +1,5 @@
 package com.example.appkhachhang.Fragment;
 
-import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -11,12 +9,10 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.util.Log;
-import android.view.ContextMenu;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -26,14 +22,17 @@ import android.view.ViewGroup;
 import android.widget.Toast;
 
 import com.example.appkhachhang.Adapter.HangSanXuatAdapter;
-import com.example.appkhachhang.Adapter.SanPhamAdapter;
+import com.example.appkhachhang.Adapter.ChiTietDienThoatAdapter;
 import com.example.appkhachhang.Adapter.SanPhamHotAdapter;
+import com.example.appkhachhang.Api.ChiTietSanPham_API;
 import com.example.appkhachhang.Api.HangSanXuat_API;
 import com.example.appkhachhang.Api.SanPham_API;
 import com.example.appkhachhang.Api.ThongKe_API;
 import com.example.appkhachhang.Interface.OnItemClickListenerHang;
 import com.example.appkhachhang.Interface.OnItemClickListenerSanPham;
 import com.example.appkhachhang.Interface.OnItemClickListenerSanPhamHot;
+import com.example.appkhachhang.Model.ChiTietDienThoai;
+import com.example.appkhachhang.Model.ChiTietGioHang;
 import com.example.appkhachhang.Model.HangSanXuat;
 import com.example.appkhachhang.Model.SanPham;
 import com.example.appkhachhang.Model.SanPhamHot;
@@ -48,10 +47,10 @@ import retrofit2.Response;
 
 public class HomeFragment extends Fragment implements OnItemClickListenerSanPham, OnItemClickListenerSanPhamHot, OnItemClickListenerHang {
     RecyclerView recyclerViewSP, recyclerViewSPHot, recyclerViewHang;
-    SanPhamAdapter sanPhamAdapter;
+    ChiTietDienThoatAdapter chiTietDienThoatAdapter;
     SanPhamHotAdapter sanPhamHotAdapter;
     HangSanXuatAdapter hangSanXuatAdapter;
-    List<SanPham> list;
+    List<ChiTietDienThoai> list;
     List<SanPhamHot> listSPHot;
     List<HangSanXuat> listHang;
     Toolbar toolbar;
@@ -89,8 +88,8 @@ public class HomeFragment extends Fragment implements OnItemClickListenerSanPham
         recyclerViewSP.setLayoutManager(linearLayoutManager);
         list = new ArrayList<>();
         getListSanPham();
-        sanPhamAdapter = new SanPhamAdapter(getContext(), list, this);
-        recyclerViewSP.setAdapter(sanPhamAdapter);
+        chiTietDienThoatAdapter = new ChiTietDienThoatAdapter(getContext(), list, this);
+        recyclerViewSP.setAdapter(chiTietDienThoatAdapter);
     }
 
     void sanPhamHot(){
@@ -149,18 +148,17 @@ public class HomeFragment extends Fragment implements OnItemClickListenerSanPham
 
 
     void getListSanPham(){
-        SanPham_API.sanPhamApi.getAllSanPham().enqueue(new Callback<List<SanPham>>() {
+        ChiTietSanPham_API.chiTietSanPhamApi.getChiTiet().enqueue(new Callback<List<ChiTietDienThoai>>() {
             @Override
-            public void onResponse(Call<List<SanPham>> call, Response<List<SanPham>> response) {
+            public void onResponse(Call<List<ChiTietDienThoai>> call, Response<List<ChiTietDienThoai>> response) {
                 list.clear();
                 list.addAll(response.body());
-                sanPhamAdapter.notifyDataSetChanged();
+                chiTietDienThoatAdapter.notifyDataSetChanged();
             }
 
             @Override
-            public void onFailure(Call<List<SanPham>> call, Throwable t) {
-                Toast.makeText(getActivity(), "Call API error: "  + t.getMessage(), Toast.LENGTH_SHORT).show();
-
+            public void onFailure(Call<List<ChiTietDienThoai>> call, Throwable t) {
+                Log.e("errorrr", "onFailure: " + t.getMessage() );
             }
         });
     }
