+++ AppKhachHang/app/src/main/java/com/example/appkhachhang/MainActivity.java
@@ -38,8 +38,8 @@ public class MainActivity extends AppCompatActivity {
     setContentView(R.layout.activity_main);
     bottomNavigationView = findViewById(R.id.bottomNavView);
     frameLayout = findViewById(R.id.frameLayout);
-    toolbar = findViewById(R.id.main_toolBar);
-    setSupportActionBar(toolbar);
+//    toolbar = findViewById(R.id.main_toolBar);
+//    setSupportActionBar(toolbar);
     bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
       @Override
       public boolean onNavigationItemSelected(@NonNull MenuItem item) {
